@@ -1,29 +1,45 @@
 import React, {Component} from 'react';
-import {
-  Text,
-  StyleSheet,
-  ScrollView,
-  View
-} from 'react-native';
+import {StyleSheet, View, ScrollView, Text} from 'react-native';
 import {Calendar} from 'react-native-calendars';
+
 
 export default class CalendarsScreen extends Component {
   constructor(props) {
     super(props);
-    this.state = {};
-    this.onDayPress = this.onDayPress.bind(this);
+    
+    this.state = {
+      selected: undefined
+    };
+  }
+
+  onDayPress = (day) => {
+    this.setState({selected: day.dateString});
   }
 
   render() {
     return (
-      <ScrollView style={styles.container}>
-        <Text style={styles.text}>Calendar with selectable date and arrows</Text>
+      <ScrollView showsVerticalScrollIndicator={false}>
+        <Text style={styles.text}>Calendar with selectable date</Text>
         <Calendar
-          onDayPress={this.onDayPress}
           style={styles.calendar}
           hideExtraDays
-          markedDates={{[this.state.selected]: {selected: true, disableTouchEvent: true, selectedDotColor: 'orange'}}}
+          onDayPress={this.onDayPress}
+          markedDates={{
+            [this.state.selected]: {
+              selected: true, 
+              disableTouchEvent: true, 
+              selectedDotColor: 'orange'
+            }
+          }}
         />
+
+        <Text style={styles.text}>Calendar with week numbers</Text>
+        <Calendar
+          style={styles.calendar}
+          hideExtraDays
+          showWeekNumbers
+        />
+
         <Text style={styles.text}>Calendar with marked dates and hidden arrows</Text>
         <Calendar
           style={styles.calendar}
@@ -32,25 +48,19 @@ export default class CalendarsScreen extends Component {
           maxDate={'2012-05-29'}
           firstDay={1}
           markedDates={{
-            '2012-05-23': {selected: true, marked: true},
-            '2012-05-24': {selected: true, marked: true, dotColor: 'green'},
+            '2012-05-23': {selected: true, marked: true, disableTouchEvent: true},
+            '2012-05-24': {selected: true, marked: true, dotColor: 'red'},
             '2012-05-25': {marked: true, dotColor: 'red'},
             '2012-05-26': {marked: true},
             '2012-05-27': {disabled: true, activeOpacity: 0}
           }}
-          // disabledByDefault={true}
           hideArrows={true}
+          // disabledByDefault={true}
         />
-        <Text style={styles.text}>Calendar with custom day component</Text>
-        <Calendar
-          style={[styles.calendar, {height: 300}]}
-          dayComponent={({date, state}) => {
-            return (<View><Text style={{textAlign: 'center', color: state === 'disabled' ? 'gray' : 'black'}}>{date.day}</Text></View>);
-          }}
-        />
+        
         <Text style={styles.text}>Calendar with period marking and spinner</Text>
         <Calendar
-          style={styles.calendar}
+          // style={styles.calendar}
           current={'2012-05-16'}
           minDate={'2012-05-10'}
           displayLoadingIndicator
@@ -76,27 +86,40 @@ export default class CalendarsScreen extends Component {
           }}
           markedDates={{
             '2012-05-17': {disabled: true},
-            '2012-05-08': {textColor: '#666'},
-            '2012-05-09': {textColor: '#666'},
-            '2012-05-14': {startingDay: true, color: 'blue', endingDay: true},
-            '2012-05-21': {startingDay: true, color: 'blue'},
+            '2012-05-08': {textColor: 'pink'},
+            '2012-05-09': {textColor: 'pink'},
+            '2012-05-14': {startingDay: true, color: 'green', endingDay: true},
+            '2012-05-21': {startingDay: true, color: 'green'},
             '2012-05-22': {endingDay: true, color: 'gray'},
             '2012-05-24': {startingDay: true, color: 'gray'},
             '2012-05-25': {color: 'gray'},
-            '2012-05-26': {endingDay: true, color: 'gray'}}}
-          hideArrows={false}
+            '2012-05-26': {endingDay: true, color: 'gray'}
+          }}
         />
+
         <Text style={styles.text}>Calendar with multi-dot marking</Text>
         <Calendar
           style={styles.calendar}
           current={'2012-05-16'}
           markingType={'multi-dot'}
           markedDates={{
-            '2012-05-08': {dots: [{key: 'vacation', color: 'blue', selectedDotColor: 'white'}, {key: 'massage', color: 'red', selectedDotColor: 'white'}], selected: true},
-            '2012-05-09': {dots: [{key: 'vacation', color: 'blue', selectedDotColor: 'red'}, {key: 'massage', color: 'red', selectedDotColor: 'blue'}], disabled: true}
+            '2012-05-08': {
+              selected: true,
+              dots: [
+                {key: 'vacation', color: 'blue', selectedDotColor: 'white'}, 
+                {key: 'massage', color: 'red', selectedDotColor: 'white'}
+              ]
+            },
+            '2012-05-09': {
+              disabled: true,
+              dots: [
+                {key: 'vacation', color: 'green', selectedDotColor: 'red'}, 
+                {key: 'massage', color: 'red', selectedDotColor: 'green'}
+              ]
+            }
           }}
-          hideArrows={false}
         />
+
         <Text style={styles.text}>Calendar with multi-period marking</Text>
         <Calendar
           style={styles.calendar}
@@ -105,35 +128,27 @@ export default class CalendarsScreen extends Component {
           markedDates={{
             '2012-05-16': {
               periods: [
-                { startingDay: true, endingDay: false, color: '#5f9ea0' },
-                { startingDay: true, endingDay: false, color: '#ffa500' },
+                {startingDay: true, endingDay: false, color: 'green'},
+                {startingDay: true, endingDay: false, color: 'orange'}
               ]
             },
             '2012-05-17': {
               periods: [
-                { startingDay: false, endingDay: true, color: '#5f9ea0' },
-                { startingDay: false, endingDay: true, color: '#ffa500' },
-                { startingDay: true, endingDay: false, color: '#f0e68c' },
+                {startingDay: false, endingDay: true, color: 'green'},
+                {startingDay: false, endingDay: true, color: 'orange'},
+                {startingDay: true, endingDay: false, color: 'pink'}
               ]
             },
             '2012-05-18': {
               periods: [
-                { startingDay: true, endingDay: true, color: '#ffa500' },
-                { color: 'transparent' },
-                { startingDay: false, endingDay: false, color: '#f0e68c' },
+                {startingDay: true, endingDay: true, color: 'orange'},
+                {color: 'transparent'},
+                {startingDay: false, endingDay: false, color: 'pink'}
               ]
-            },
+            }
           }}
-          hideArrows={false}
         />
-        <Text style={styles.text}>Calendar with week numbers</Text>
-        <Calendar
-          onDayPress={this.onDayPress}
-          style={styles.calendar}
-          hideExtraDays
-          showWeekNumbers
-          markedDates={{[this.state.selected]: {selected: true}}}
-        />
+        
         <Text style={styles.text}>Custom calendar with custom marking type</Text>
         <Calendar
           style={styles.calendar}
@@ -150,32 +165,33 @@ export default class CalendarsScreen extends Component {
                   elevation: 2
                 },
                 text: {
-                  color: 'blue',
-                },
+                  color: 'red'
+                }
               }
             },
-            '2018-03-08': {selected: true},
+            '2018-03-08': {
+              selected: true
+            },
             '2018-03-09': {
               customStyles: {
                 container: {
                   backgroundColor: 'red',
-                  elevation: 4,
+                  elevation: 4
                 },
                 text: {
-                  color: 'white',
-                },
+                  color: 'white'
+                }
               }
             },
-            '2018-03-10': {disabled: true},
             '2018-03-14': {
               customStyles: {
                 container: {
-                  backgroundColor: 'green',
+                  backgroundColor: 'green'
                 },
                 text: {
-                  color: 'white',
-                },
-              },
+                  color: 'white'
+                }
+              }
             },
             '2018-03-15': {
               customStyles: {
@@ -184,99 +200,80 @@ export default class CalendarsScreen extends Component {
                   elevation: 2
                 },
                 text: {
-                  color: 'yellow',
-                },
+                  color: 'yellow'
+                }
               }
             },
-            '2018-03-20': {
-              customStyles: {
-                container: {
-                  backgroundColor: 'pink',
-                  elevation: 4,
-                },
-                text: {
-                  color: 'blue',
-                },
-              }
+            '2018-03-21': {
+              disabled: true
             },
-            '2018-03-21': {disabled: true},
             '2018-03-28': {
               customStyles: {
-                container: {
-                  backgroundColor: 'green',
-                },
                 text: {
                   color: 'black',
                   fontWeight: 'bold'
-                },
-              },
-            },
-            '2018-03-29': {
-              customStyles: {
-                container: {
-                  backgroundColor: 'white',
-                  elevation: 2
-                },
-                text: {
-                  color: 'blue',
-                },
+                }
               }
             },
             '2018-03-30': {
               customStyles: {
                 container: {
-                  backgroundColor: 'violet',
+                  backgroundColor: 'pink',
                   elevation: 4,
-                  borderColor: 'red',
-                  borderWidth: 5,
+                  borderColor: 'purple',
+                  borderWidth: 5
                 },
                 text: {
                   marginTop: 3,
                   fontSize: 11,
-                  color: 'yellow',
-                },
+                  color: 'black'
+                }
               }
             },
             '2018-03-31': {
               customStyles: {
                 container: {
-                  backgroundColor: 'green',
-                  borderRadius: 0,
-                },
-                text: {
-                  color: 'white',
-                },
-              },
-            }}}
-          hideArrows={false}
+                  backgroundColor: 'orange',
+                  borderRadius: 0
+                }
+              }
+            }
+          }}
+        />
+
+        <Text style={styles.text}>Calendar with custom day component</Text>
+        <Calendar
+          style={[
+            styles.calendar,
+            {
+              height: 250, 
+              borderBottomWidth: 1, 
+              borderBottomColor: 'lightgrey'
+            }
+          ]}
+          dayComponent={({date, state}) => {
+            return (
+              <View>
+                <Text style={{textAlign: 'center', color: state === 'disabled' ? 'gray' : 'black'}}>
+                  {date.day}
+                </Text>
+              </View>
+            );
+          }}
         />
       </ScrollView>
     );
-  }
-
-  onDayPress(day) {
-    this.setState({
-      selected: day.dateString
-    });
   }
 }
 
 const styles = StyleSheet.create({
   calendar: {
-    borderTopWidth: 1,
-    paddingTop: 5,
-    borderBottomWidth: 1,
-    borderColor: '#eee',
-    height: 350
+    marginBottom: 10
   },
   text: {
     textAlign: 'center',
-    borderColor: '#bbb',
     padding: 10,
-    backgroundColor: '#eee'
-  },
-  container: {
-    flex: 1,
-    backgroundColor: 'gray'
+    backgroundColor: 'lightgrey',
+    fontSize: 16
   }
 });
