@@ -1,15 +1,12 @@
 import React, {Component} from 'react';
-
 import {CalendarList} from 'react-native-calendars';
 
-export default class CalendarsList extends Component {
-  constructor(props) {
-    super(props);
-  }
 
+export default class CalendarsList extends Component {
+  
   render() {
     return (
-      <CalendarList current={'2012-05-16'} pastScrollRange={24} futureScrollRange={24} />
+      <CalendarList current={'2012-05-16'} pastScrollRange={24} futureScrollRange={24}/>
     );
   }
 }
