@@ -9,13 +9,15 @@ import {
   TouchableOpacity,
   Button
 } from 'react-native';
-import {ExpandableCalendar, AgendaList, CalendarProvider} from 'react-native-calendars';
+import {ExpandableCalendar, AgendaList, CalendarProvider, WeekCalendar} from 'react-native-calendars';
 
 
 const today = new Date().toISOString().split('T')[0];
 const fastDate = getPastDate(3); 
 const futureDates = getFutureDates(9);
 const dates = [fastDate, today].concat(futureDates);
+const themeColor = '#00AAAF';
+const lightThemeColor = '#EBF9F9';
 
 function getFutureDates(days) {
   const array = [];
@@ -88,7 +90,7 @@ export default class ExpandableCalendarScreen extends Component {
         </View>
         <Text style={styles.itemTitleText}>{item.title}</Text>
         <View style={styles.itemButtonContainer}>
-          <Button title={'Info'} onPress={this.buttonPressed}/>
+          <Button color={'grey'} title={'Info'} onPress={this.buttonPressed}/>
         </View>
       </TouchableOpacity>
     );
@@ -97,7 +99,7 @@ export default class ExpandableCalendarScreen extends Component {
   getMarkedDates = () => {
     const marked = {};
     ITEMS.forEach(item => {
-      // only mark dates with data
+      // NOTE: only mark dates with data
       if (item.data && item.data.length > 0 && !_.isEmpty(item.data[0])) {
         marked[item.title] = {marked: true};
       }
@@ -106,29 +108,22 @@ export default class ExpandableCalendarScreen extends Component {
   }
 
   getTheme = () => {
-    const themeColor = '#0059ff';
-    const lightThemeColor = '#e6efff';
-    const disabledColor = '#a6acb1';
-    const black = '#20303c';
-    const white = '#ffffff';
+    const disabledColor = 'grey';
     
     return {
       // arrows
-      arrowColor: black,
+      arrowColor: 'black',
       arrowStyle: {padding: 0},
       // month
-      monthTextColor: black,
+      monthTextColor: 'black',
       textMonthFontSize: 16,
       textMonthFontFamily: 'HelveticaNeue',
       textMonthFontWeight: 'bold',
       // day names
-      textSectionTitleColor: black,
+      textSectionTitleColor: 'black',
       textDayHeaderFontSize: 12,
       textDayHeaderFontFamily: 'HelveticaNeue',
       textDayHeaderFontWeight: 'normal',
-      // today
-      todayBackgroundColor: lightThemeColor,
-      todayTextColor: themeColor,
       // dates
       dayTextColor: themeColor,
       textDayFontSize: 18,
@@ -137,12 +132,12 @@ export default class ExpandableCalendarScreen extends Component {
       textDayStyle: {marginTop: Platform.OS === 'android' ? 2 : 4},
       // selected date
       selectedDayBackgroundColor: themeColor,
-      selectedDayTextColor: white,
+      selectedDayTextColor: 'white',
       // disabled date
       textDisabledColor: disabledColor,
       // dot (marked date)
       dotColor: themeColor,
-      selectedDotColor: white,
+      selectedDotColor: 'white',
       disabledDotColor: disabledColor,
       dotStyle: {marginTop: -2}
     };
@@ -154,25 +149,34 @@ export default class ExpandableCalendarScreen extends Component {
         date={ITEMS[0].title} 
         onDateChanged={this.onDateChanged} 
         onMonthChange={this.onMonthChange}
-        theme={{todayButtonTextColor: '#0059ff'}} 
         showTodayButton 
         disabledOpacity={0.6}
+        // theme={{
+        //   todayButtonTextColor: themeColor
+        // }} 
         // todayBottomMargin={16}
-      >
-        <ExpandableCalendar 
-          // horizontal={false}
-          // hideArrows
-          // disablePan
-          // hideKnob
-          // initialPosition={ExpandableCalendar.positions.OPEN}
-          firstDay={1}
-          markedDates={this.getMarkedDates()} // {'2019-06-01': {marked: true}, '2019-06-02': {marked: true}, '2019-06-03': {marked: true}};
-          theme={this.getTheme()}
-          leftArrowImageSource={require('../img/previous.png')}
-          rightArrowImageSource={require('../img/next.png')}
-          // calendarStyle={styles.calendar}
-          // headerStyle={styles.calendar} // for horizontal only
-        />
+      > 
+        {this.props.weekView ? 
+          <WeekCalendar
+            firstDay={1}
+            markedDates={this.getMarkedDates()}
+          /> :
+          <ExpandableCalendar 
+            // horizontal={false}
+            // hideArrows
+            // disablePan
+            // hideKnob
+            // initialPosition={ExpandableCalendar.positions.OPEN}
+            // calendarStyle={styles.calendar}
+            // headerStyle={styles.calendar} // for horizontal only
+            // disableWeekScroll
+            // theme={this.getTheme()}
+            firstDay={1}
+            markedDates={this.getMarkedDates()} // {'2019-06-01': {marked: true}, '2019-06-02': {marked: true}, '2019-06-03': {marked: true}};
+            leftArrowImageSource={require('../img/previous.png')}
+            rightArrowImageSource={require('../img/next.png')}
+          />
+        }
         <AgendaList
           sections={ITEMS}
           extraData={this.state}
@@ -190,14 +194,15 @@ const styles = StyleSheet.create({
     paddingRight: 20
   },
   section: {
-    backgroundColor: '#f0f4f7', 
-    color: '#79838a'
+    backgroundColor: lightThemeColor, 
+    color: 'grey',
+    textTransform: 'capitalize'
   },
   item: {
     padding: 20, 
     backgroundColor: 'white', 
     borderBottomWidth: 1, 
-    borderBottomColor: '#e8ecf0', 
+    borderBottomColor: 'lightgrey', 
     flexDirection: 'row'
   },
   itemHourText: {
@@ -224,10 +229,10 @@ const styles = StyleSheet.create({
     height: 52, 
     justifyContent: 'center',
     borderBottomWidth: 1, 
-    borderBottomColor: '#e8ecf0' 
+    borderBottomColor: 'lightgrey' 
   },
   emptyItemText: {
-    color: '#79838a',
+    color: 'lightgrey',
     fontSize: 14
   }
 });
