@@ -1,18 +1,16 @@
 import React, {Component} from 'react';
-import {
-  Text,
-  View,
-  TouchableOpacity,
-  StyleSheet,
-} from 'react-native';
+import {Platform, StyleSheet, View, TouchableOpacity, Text, Image} from 'react-native';
 import {Navigation} from 'react-native-navigation';
 
+
+const appIcon = require('../img/app-icon-120x120.png');
 
 export default class MenuScreen extends Component {
 
   render() {
     return (
-      <View>
+      <View style={styles.container}>
+        <Image source={appIcon} style={styles.image}/>
         <TouchableOpacity style={styles.menu} onPress={this.onCalendarsPress.bind(this)}>
           <Text style={styles.menuText}>Calendars</Text>
         </TouchableOpacity>
@@ -28,18 +26,27 @@ export default class MenuScreen extends Component {
         <TouchableOpacity style={styles.menu} onPress={this.onExpandablePress.bind(this)}>
           <Text style={styles.menuText}>Expandable Calendar</Text>
         </TouchableOpacity>
+        <TouchableOpacity style={styles.menu} onPress={this.onWeekPress.bind(this)}>
+          <Text style={styles.menuText}>Week Calendar</Text>
+        </TouchableOpacity>
       </View>
     );
   }
 
-  pushScreen(screen) {
+  pushScreen(screen, props) {
     Navigation.push(this.props.componentId, {
       component: {
         name: screen,
+        passProps: props,
         options: {
           topBar: {
             title: {
               text: screen
+            },
+            backButton: {
+              accessibilityLabel: 'back',
+              showTitle: false, // iOS only
+              color: Platform.OS === 'ios' ? '#2d4150' : undefined
             }
           }
         }
@@ -66,16 +73,34 @@ export default class MenuScreen extends Component {
   onExpandablePress() {
     this.pushScreen('ExpandableCalendar');
   }
+
+  onWeekPress() {
+    this.pushScreen('ExpandableCalendar', {weekView: true});
+  }
 }
 
 const styles = StyleSheet.create({
+  container: {
+    flex: 1, 
+    alignItems: 'center'
+  },
+  image: {
+    margin: 30,
+    width: 90,
+    height: 90
+  },
   menu: {
-    height: 50,
-    justifyContent: 'center',
-    paddingLeft: 15,
-    borderBottomWidth: 1
+    width: 300,
+    padding: 10,
+    margin: 10,
+    // backgroundColor: '#f2F4f5',
+    alignItems: 'center',
+    borderRadius: 20,
+    borderWidth: 1,
+    borderColor: '#7a92a5'
   },
   menuText: {
-    fontSize: 18
+    fontSize: 18,
+    color: '#2d4150'
   }
 });
