@@ -14,6 +14,7 @@ class CalendarListItem extends Component {
 
   constructor(props) {
     super(props);
+
     this.style = styleConstructor(props.theme);
   }
 
@@ -29,7 +30,14 @@ class CalendarListItem extends Component {
     if (this.props.onPressArrowLeft) {
       this.props.onPressArrowLeft(_, monthClone);
     } else if (this.props.scrollToMonth) {
+      const currentMonth = monthClone.getMonth();
       monthClone.addMonths(-1);
+
+      // Make sure we actually get the previous month, not just 30 days before currentMonth.
+      while (monthClone.getMonth() === currentMonth) {
+        monthClone.setDate(monthClone.getDate() - 1);
+      }
+
       this.props.scrollToMonth(monthClone);
     }
   }
@@ -74,7 +82,10 @@ class CalendarListItem extends Component {
           onPressArrowLeft={this.props.horizontal ? this.onPressArrowLeft : this.props.onPressArrowLeft}
           onPressArrowRight={this.props.horizontal ? this.onPressArrowRight : this.props.onPressArrowRight}
           headerStyle={this.props.horizontal ? this.props.headerStyle : undefined}
-        />);
+          accessibilityElementsHidden={this.props.accessibilityElementsHidden} // iOS
+          importantForAccessibility={this.props.importantForAccessibility} // Android
+        />
+      );
     } else {
       const text = row.toString();
 
