@@ -1,9 +1,10 @@
-import React, { Component } from 'react';
-import { TouchableOpacity, Text, View } from 'react-native';
+import React, {Component} from 'react';
+import {TouchableOpacity, Text, View} from 'react-native';
 import PropTypes from 'prop-types';
 import {shouldUpdate} from '../../../component-updater';
 
 import styleConstructor from './style';
+
 
 class Day extends Component {
   static displayName = 'IGNORE';
@@ -11,22 +12,28 @@ class Day extends Component {
   static propTypes = {
     // TODO: disabled props should be removed
     state: PropTypes.oneOf(['disabled', 'today', '']),
-
     // Specify theme properties to override specific styles for calendar parts. Default = {}
     theme: PropTypes.object,
     marking: PropTypes.any,
     onPress: PropTypes.func,
-    date: PropTypes.object,
+    onLongPress: PropTypes.func,
+    date: PropTypes.object
   };
 
   constructor(props) {
     super(props);
+
     this.style = styleConstructor(props.theme);
+
     this.onDayPress = this.onDayPress.bind(this);
+    this.onDayLongPress = this.onDayLongPress.bind(this);
   }
 
   onDayPress() {
     this.props.onPress(this.props.date);
+  }
+  onDayLongPress() {
+    this.props.onLongPress(this.props.date);
   }
 
   shouldComponentUpdate(nextProps) {
@@ -46,24 +53,24 @@ class Day extends Component {
         const style = [
           ...baseDotStyle,
           {
-            backgroundColor: period.color,
-          },
+            backgroundColor: period.color
+          }
         ];
         if (period.startingDay) {
           style.push({
             borderTopLeftRadius: 2,
             borderBottomLeftRadius: 2,
-            marginLeft: 4,
+            marginLeft: 4
           });
         }
         if (period.endingDay) {
           style.push({
             borderTopRightRadius: 2,
             borderBottomRightRadius: 2,
-            marginRight: 4,
+            marginRight: 4
           });
         }
-        return <View key={index} style={style} />;
+        return <View key={index} style={style}/>;
       });
     }
     return;
@@ -72,37 +79,36 @@ class Day extends Component {
   render() {
     const containerStyle = [this.style.base];
     const textStyle = [this.style.text];
-
     const marking = this.props.marking || {};
     const periods = this.renderPeriods(marking);
+    const isDisabled = typeof marking.disabled !== 'undefined' ? marking.disabled : this.props.state === 'disabled';
 
     if (marking.selected) {
       containerStyle.push(this.style.selected);
       textStyle.push(this.style.selectedText);
-    } else if (
-      typeof marking.disabled !== 'undefined'
-        ? marking.disabled
-        : this.props.state === 'disabled'
-    ) {
+    } else if (isDisabled) {
       textStyle.push(this.style.disabledText);
     } else if (this.props.state === 'today') {
       containerStyle.push(this.style.today);
       textStyle.push(this.style.todayText);
     }
     return (
-      <View
-        style={{
-          alignSelf: 'stretch'
-        }}>
-        <TouchableOpacity testID={this.props.testID} style={containerStyle} onPress={this.onDayPress}>
+      <View style={{alignSelf: 'stretch'}}>
+        <TouchableOpacity 
+          testID={this.props.testID} 
+          style={containerStyle} 
+          onPress={this.onDayPress}
+          onLongPress={this.onDayLongPress}
+          disabled={marking.disableTouchEvent}
+          accessible
+          accessibilityRole={isDisabled ? undefined : 'button'}
+          accessibilityLabel={this.props.accessibilityLabel}
+        >
           <Text allowFontScaling={false} style={textStyle}>
             {String(this.props.children)}
           </Text>
         </TouchableOpacity>
-        <View
-          style={{
-            alignSelf: 'stretch',
-          }}>
+        <View style={{alignSelf: 'stretch'}}>
           {periods}
         </View>
       </View>
