@@ -1,14 +1,12 @@
 import _ from 'lodash';
 import PropTypes from 'prop-types';
 import React, {Component} from 'react';
-import {
-  TouchableWithoutFeedback,
-  Text,
-  View} from 'react-native';
+import {TouchableWithoutFeedback, Text, View} from 'react-native';
 import {shouldUpdate} from '../../../component-updater';
 
 import * as defaultStyle from '../../../style';
 import styleConstructor from './style';
+
 
 class Day extends Component {
   static displayName = 'IGNORE';
@@ -16,22 +14,21 @@ class Day extends Component {
   static propTypes = {
     // TODO: selected + disabled props should be removed
     state: PropTypes.oneOf(['selected', 'disabled', 'today', '']),
-
     // Specify theme properties to override specific styles for calendar parts. Default = {}
     theme: PropTypes.object,
     marking: PropTypes.any,
-
     onPress: PropTypes.func,
     onLongPress: PropTypes.func,
     date: PropTypes.object,
-
-    markingExists: PropTypes.bool,
+    markingExists: PropTypes.bool
   };
 
   constructor(props) {
     super(props);
+    
     this.theme = {...defaultStyle, ...(props.theme || {})};
     this.style = styleConstructor(props.theme);
+    
     this.markingStyle = this.getDrawingStyle(props.marking || []);
     this.onDayPress = this.onDayPress.bind(this);
     this.onDayLongPress = this.onDayLongPress.bind(this);
@@ -196,7 +193,12 @@ class Day extends Component {
       <TouchableWithoutFeedback
         testID={this.props.testID}
         onPress={this.onDayPress}
-        onLongPress={this.onDayLongPress}>
+        onLongPress={this.onDayLongPress}
+        disabled={this.props.marking.disableTouchEvent}
+        accessible
+        accessibilityRole={this.props.marking.disableTouchEvent ? undefined : 'button'}
+        accessibilityLabel={this.props.accessibilityLabel}
+      >
         <View style={this.style.wrapper}>
           {fillers}
           <View style={containerStyle}>
