@@ -1,8 +1,8 @@
 import {StyleSheet} from 'react-native';
 import * as defaultStyle from '../../../style';
 
-const STYLESHEET_ID = 'stylesheet.day.period';
 
+const STYLESHEET_ID = 'stylesheet.day.period';
 const FILLER_HEIGHT = 34;
 
 export default function styleConstructor(theme={}) {
@@ -39,7 +39,7 @@ export default function styleConstructor(theme={}) {
       fontSize: appStyle.textDayFontSize,
       fontFamily: appStyle.textDayFontFamily,
       fontWeight: appStyle.textDayFontWeight,
-      color: appStyle.dayTextColor || '#2d4150',
+      color: appStyle.dayTextColor,
       backgroundColor: 'rgba(255, 255, 255, 0)'
     },
     today: {
@@ -47,7 +47,7 @@ export default function styleConstructor(theme={}) {
     },
     todayText: {
       fontWeight: '500',
-      color: theme.todayTextColor || appStyle.dayTextColor,
+      color: theme.todayTextColor || appStyle.dayTextColor
       //color: appStyle.textLinkColor
     },
     disabledText: {
