@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import {ActivityIndicator} from 'react-native';
+import {ActivityIndicator, Platform} from 'react-native';
 import {View, Text, TouchableOpacity, Image} from 'react-native';
 import XDate from 'xdate';
 import PropTypes from 'prop-types';
@@ -10,7 +10,7 @@ import {CHANGE_MONTH_LEFT_ARROW, CHANGE_MONTH_RIGHT_ARROW} from '../../testIDs';
 
 class CalendarHeader extends Component {
   static displayName = 'IGNORE';
-  
+
   static propTypes = {
     theme: PropTypes.object,
     hideArrows: PropTypes.bool,
@@ -22,11 +22,15 @@ class CalendarHeader extends Component {
     hideDayNames: PropTypes.bool,
     weekNumbers: PropTypes.bool,
     onPressArrowLeft: PropTypes.func,
-    onPressArrowRight: PropTypes.func
+    onPressArrowRight: PropTypes.func,
+    disableArrowLeft: PropTypes.bool,
+    disableArrowRight: PropTypes.bool,
+    webAriaLevel: PropTypes.number
   };
 
   static defaultProps = {
-    monthFormat: 'MMMM yyyy'
+    monthFormat: 'MMMM yyyy',
+    webAriaLevel: 1
   };
 
   constructor(props) {
@@ -65,6 +69,15 @@ class CalendarHeader extends Component {
     if (nextProps.monthFormat !== this.props.monthFormat) {
       return true;
     }
+    if (nextProps.renderArrow !== this.props.renderArrow) {
+      return true;
+    }
+    if (nextProps.disableArrowLeft !== this.props.disableArrowLeft) {
+      return true;
+    }
+    if (nextProps.disableArrowRight !== this.props.disableArrowRight) {
+      return true;
+    }
     return false;
   }
 
@@ -85,8 +98,8 @@ class CalendarHeader extends Component {
   }
 
   render() {
-    let leftArrow = <View />;
-    let rightArrow = <View />;
+    let leftArrow = <View/>;
+    let rightArrow = <View/>;
     let weekDaysNames = weekDayNames(this.props.firstDay);
     const {testID} = this.props;
 
@@ -94,6 +107,7 @@ class CalendarHeader extends Component {
       leftArrow = (
         <TouchableOpacity
           onPress={this.onPressLeft}
+          disabled={this.props.disableArrowLeft}
           style={this.style.arrow}
           hitSlop={{left: 20, right: 20, top: 20, bottom: 20}}
           testID={testID ? `${CHANGE_MONTH_LEFT_ARROW}-${testID}`: CHANGE_MONTH_LEFT_ARROW}
@@ -102,13 +116,14 @@ class CalendarHeader extends Component {
             ? this.props.renderArrow('left')
             : <Image
               source={require('../img/previous.png')}
-              style={this.style.arrowImage}
+              style={this.props.disableArrowLeft ? this.style.disabledArrowImage : this.style.arrowImage}
             />}
         </TouchableOpacity>
       );
       rightArrow = (
         <TouchableOpacity
           onPress={this.onPressRight}
+          disabled={this.props.disableArrowRight}
           style={this.style.arrow}
           hitSlop={{left: 20, right: 20, top: 20, bottom: 20}}
           testID={testID ? `${CHANGE_MONTH_RIGHT_ARROW}-${testID}`: CHANGE_MONTH_RIGHT_ARROW}
@@ -117,7 +132,7 @@ class CalendarHeader extends Component {
             ? this.props.renderArrow('right')
             : <Image
               source={require('../img/next.png')}
-              style={this.style.arrowImage}
+              style={this.props.disableArrowRight ? this.style.disabledArrowImage : this.style.arrowImage}
             />}
         </TouchableOpacity>
       );
@@ -128,30 +143,49 @@ class CalendarHeader extends Component {
       indicator = <ActivityIndicator color={this.props.theme && this.props.theme.indicatorColor}/>;
     }
 
+    const webProps = Platform.OS === 'web' ? {'aria-level': this.props.webAriaLevel} : {};
+
     return (
-      <View style={this.props.style}>
+      <View 
+        style={this.props.style} 
+        accessible
+        accessibilityRole={'adjustable'}
+        accessibilityActions={[
+          {name: 'increment', label: 'increment'}, 
+          {name: 'decrement', label: 'decrement'}
+        ]}
+        onAccessibilityAction={this.onAccessibilityAction}
+        accessibilityElementsHidden={this.props.accessibilityElementsHidden} // iOS
+        importantForAccessibility={this.props.importantForAccessibility} // Android
+      >
         <View style={this.style.header}>
           {leftArrow}
-          <View style={{ flexDirection: 'row' }}>
-            <Text allowFontScaling={false} style={this.style.monthText} accessibilityTraits='header'>
+          <View style={{flexDirection: 'row'}}>
+            <Text
+              allowFontScaling={false}
+              style={this.style.monthText}
+              {...webProps}
+            >
               {this.props.month.toString(this.props.monthFormat)}
             </Text>
             {indicator}
           </View>
           {rightArrow}
         </View>
-        {
-          !this.props.hideDayNames &&
+        {!this.props.hideDayNames &&
           <View style={this.style.week}>
-            {this.props.weekNumbers && <Text allowFontScaling={false} style={this.style.dayHeader}></Text>}
+            {this.props.weekNumbers && 
+              <Text allowFontScaling={false} style={this.style.dayHeader}></Text>
+            }
             {weekDaysNames.map((day, idx) => (
-              <Text 
-                allowFontScaling={false} 
-                key={idx} 
-                accessible={false} 
-                style={this.style.dayHeader} 
-                numberOfLines={1} 
-                importantForAccessibility='no'
+              <Text
+                allowFontScaling={false}
+                key={idx}
+                style={this.style.dayHeader}
+                numberOfLines={1}
+                accessibilityLabel={''}
+                // accessible={false} // not working
+                // importantForAccessibility='no'
               >
                 {day}
               </Text>
@@ -160,6 +194,19 @@ class CalendarHeader extends Component {
         }
       </View>
     );
+  }
+
+  onAccessibilityAction = event => {
+    switch (event.nativeEvent.actionName) {
+    case 'decrement':
+      this.onPressLeft();
+      break;
+    case 'increment':
+      this.onPressRight();
+      break;
+    default:
+      break;
+    }
   }
 }
 
