@@ -29,7 +29,8 @@ class AgendaList extends Component {
   }
 
   static defaultProps = {
-    dayFormat: 'dddd, MMM d'
+    dayFormat: 'dddd, MMM d',
+    stickySectionHeadersEnabled: true
   }
 
   constructor(props) {
@@ -120,17 +121,19 @@ class AgendaList extends Component {
   }
 
   renderSectionHeader = ({section: {title}}) => {
-    const today = XDate().toString(this.props.dayFormat).toUpperCase();
-    const date = XDate(title).toString(this.props.dayFormat).toUpperCase();
+    const today = XDate().toString(this.props.dayFormat);
+    const date = XDate(title).toString(this.props.dayFormat);
     const todayString = XDate.locales[XDate.defaultLocale].today || commons.todayString;
-    const sectionTitle = date === today ? `${todayString.toUpperCase()}, ${date}` : date;
+    const sectionTitle = date === today ? `${todayString}, ${date}` : date;
     
     return (
       <Text allowFontScaling={false} style={[this.style.sectionText, this.props.sectionStyle]} onLayout={this.onHeaderLayout}>{sectionTitle}</Text>
     );
   }
 
-  keyExtractor = (item, index) => String(index);
+  keyExtractor = (item, index) => {
+    return _.isFunction(this.props.keyExtractor) ? this.props.keyExtractor(item, index) : String(index);
+  }
 
   render() {
     return (
@@ -139,7 +142,6 @@ class AgendaList extends Component {
         ref={this.list}
         keyExtractor={this.keyExtractor}
         showsVerticalScrollIndicator={false}
-        stickySectionHeadersEnabled
         onViewableItemsChanged={this.onViewableItemsChanged}
         viewabilityConfig={this.viewabilityConfig}
         renderSectionHeader={this.renderSectionHeader}
