@@ -1,17 +1,15 @@
-import { Platform } from 'react-native';
+export const foregroundColor = 'white';
+export const backgroundColor = '#f2F4f5'; // Agenda's reservations bg color
 
-export const foregroundColor = '#ffffff';
-export const backgroundColor = '#f4f4f4'; // Agenda's reservations bg color
-export const separatorColor = '#e8e9ec'; // not in use
-
-export const processedColor = '#a7e0a3';
-export const processingColor = '#ffce5c';
-export const failedColor = '#f67e7e';
+export const separatorColor = '#f2F4f5'; // not in use
+export const processedColor = '#a7e0a3'; // not in use
+export const processingColor = '#ffce5c'; // not in use
+export const failedColor = '#f67e7e'; // not in use
+export const textSecondaryColor = '#7a92a5'; // not in use
 
 export const textDefaultColor = '#2d4150';
 export const textColor = '#43515c';
-export const textLinkColor = '#00adf5';
-export const textSecondaryColor = '#7a92a5';
+export const textLinkColor = '#00BBF2';
 
 export const textDayFontFamily = 'System';
 export const textMonthFontFamily = 'System';
@@ -45,11 +43,12 @@ export const selectedDotColor = foregroundColor;
 export const disabledDotColor = undefined;
 export const todayDotColor = undefined;
 export const arrowColor = textLinkColor;
+export const disabledArrowColor = '#d9e1e8';
 export const monthTextColor = textDefaultColor;
 export const indicatorColor = undefined; // use the default color of React Native ActivityIndicator
 export const agendaDayTextColor = '#7a92a5';
 export const agendaDayNumColor = '#7a92a5';
 export const agendaTodayColor = textLinkColor;
-export const agendaKnobColor = Platform.OS === 'ios' ? '#f2F4f5' : '#4ac4f7';
+export const agendaKnobColor = '#f2F4f5';
 export const todayButtonTextColor = textLinkColor;
-export const todayButtonPosition = undefined; // right' / 'left' (default)
+export const todayButtonPosition = undefined; // right' / 'left'(default)
